fix: make episodeId type Int

+++ domain/src/main/java/com/amsterdam/domain/useCase/home/GetEpisodeVideosByEpisodeId.kt
@@ -8,7 +8,7 @@ class GetEpisodeVideosByEpisodeId (
     suspend operator fun invoke(
         tvShowId: Long,
         seasonNumber: Int,
-        episodeNumber: Long
-    ) = tvShowRepository.getEpisodeVideoUrl(tvShowId, seasonNumber, episodeNumber.toInt())
+        episodeNumber: Int
+    ) = tvShowRepository.getEpisodeVideoUrl(tvShowId, seasonNumber, episodeNumber)
 
 }
